Split cpp codegen scheduling and outer loop fusion
